fix: FAILED to fix Registration due to poor design
I tried to fix why the new User Registration Does not save to the Local Storage.
It seemed Complicated as index.js, register.js, and login.js has dependencies.
it is my fault but this Project I developed in 1st Year at College so it is messy.
so maybe I come later to re-engineer it for the sake of revising Good Memories. (Failed)